Register oink's CuTeDSL LayerNorm kernel as a CUDA dispatch override for `aten::native_layer_norm` via `torch._native`. On SM100 (Blackwell) devices with CuTeDSL installed, the oink kernel handles 2D bf16/fp16 inputs; otherwise falls back to the original CUDA kernel.                         
   
  - `layernorm.py`: Registration + gating logic. The adapter maps between aten's `(out, mean, rstd)` return order and oink's `(out, rstd, mean)` return order.                                                                                                                                        
  - `layernorm_kernels.py`: CuTeDSL SM100 LayerNorm forward (2-stage: mean pass then variance/normalization) and backward (separate dx and dweight/dbias kernels).
  - Tests added to `test_oink_kernels.py` for gating, registration, and correctness.   


[ghstack-poisoned]

+++ test/python_native/test_oink_kernels.py
@@ -267,9 +267,7 @@ class TestOinkLayernormCorrectness(TestCase):
             b = torch.randn(N, device="cuda", dtype=torch.float32)
             eps = 1e-5
 
-            ref = torch.nn.functional.layer_norm(x.float(), [N], w, b, eps).to(
-                x.dtype
-            )
+            ref = torch.nn.functional.layer_norm(x.float(), [N], w, b, eps).to(x.dtype)
             result, _, _ = torch.native_layer_norm(x, [N], w, b, eps)
             torch.testing.assert_close(
                 result,
